feat(command): Separate the get and set commands into different files.
